move unsupported_instruction test to gear-core
